refactor(hotkey): move hotkey settings logic to component
